updated code

+++ backend/src/users/users.controller.ts
@@ -9,21 +9,31 @@ import {
   UseGuards,
   Request,
   UnauthorizedException,
+  BadRequestException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { JWTAuthGuard } from 'src/auth/guards/jwt-auth.guard';
-
+import { Role } from 'src/auth/enums/role.enum';
+import { Roles } from 'src/auth/decorators/roles.decorator';
+@UseGuards(JWTAuthGuard)
 @Controller('users')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
+
+  async getUserExist(id: number){
+    const user = await this.usersService.findOne(id)
+    if(!user) throw new BadRequestException('No such user exist')
+   return user
+  }
 
   @Post()
   create(@Body() createUserDto: CreateUserDto) {
     return this.usersService.create(createUserDto);
   }
-
+ 
+  @Roles(Role.ADMIN)
   @Get()
   findAll() {
     return this.usersService.findAll();
@@ -39,23 +49,35 @@ export class UsersController {
     return this.usersService.findByEmail(email);
   }
   
-@UseGuards(JWTAuthGuard)
-@Get('profile')
-async getProfile(@Request() req) {
-  const userId = req.user?.id;
-  if (!userId) throw new UnauthorizedException('User ID missing');
+ 
+  @Get('profile')
+  async getProfile(@Request() req) {
+    const userId = req.user?.id;
+    if (!userId) throw new UnauthorizedException('User ID missing');
 
-  const user = await this.usersService.findOne(userId); // ✅ userId is number
-  return user;
-}
-
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.usersService.update(+id, updateUserDto);
+    const user = await this.getUserExist(userId)
+    return user
   }
 
+
+  
+  @Patch(':id')
+  async update(@Param('id') id: string , @Body() updateUserDto: UpdateUserDto) {
+     //does user exist
+     const user = await this.getUserExist(+id)
+     if (user){
+       return this.usersService.update(+id, updateUserDto);
+
+     }
+  }
+  
+  @Roles(Role.ADMIN)
   @Delete(':id')
-  deleteUser(@Param('id') id: string) {
+  async deleteUser(@Param('id') id: string) {
+    const user = await this.getUserExist(+id)
+    if(user){
     return this.usersService.deleteUser(+id);
+      
+    }
   }
 }
